Implement media doesFileExists().
Current way to do that is to provide, and if t
fails, it doesnt exists, but if it succeeds, then
you do a transfer :-( So doesFileExists to the rescue

The current implementation does a partial
transfer of 100 bytes, as using NOBODY option
of curl doesn't work for ftp.

It can be optimized using different methods
for different URL schemes. It is not yet
used but it wll be used in the source 
probing.

Marius, please give it a look when you are back.