fix build after for_ was changed to use auto.
the construct for_(i, 0UL, v.size()) did not build becasue
size_t is not 0UL on all systems.

Replacing 0 with a size_t cast or the container size_type would
defeat the purpose of using a for_ construct.

For iterations where the index is needed a classic for is shorter.

Also the question remains open whether for_ is needed now that C++11
provides the for (auto x : container) construct.