Add an "--ignore-unknown" global option to zypper.
If zypper is given this option and cannot find a package, it will neither
return an error code greater than 0 nor fail to process other packages. This is
especially useful when running zypper from a shell script with the
"--non-interactive" global option.

This currently only applies to the "install", "remove", "source install" and
"update" commands, but could easily be extended to other commands.