Fix mixing C and C++ in templates
This kills the favorite error for some gcc versions about C in the
templates.

*snip*
/usr/lib/gcc/x86_64-pc-linux-gnu/4.6.3/include/g++-v4/bits/basic_string.tcc:1068:3:
error: template with C linkage
make[2]: *** [zypp/CMakeFiles/zypp.dir/parser/xml/Node.cc.o] Error 1
*snip*

Basically we don't need to put this stuff here as it is done in the
headers themselves, so just prune the problematic code.