Drop package manager script
It is deprecated as original reason in BNC#222757 is no longer valid.
It is also really bad design decision to let decide library which its
GUI user should be used. If someone miss it then add it to top level
package and not in library. Suse part is also no longer valid due to
FATE#316458 which removes yast -i and suggest zypper usage.