SUSEMediaVerifier: add debug logs
isDesiredMedia() media may fail, which ultimately causes a
MediaNotDesiredException to be thrown. To ease debugging, print the
values of the (cached) media id and vendor in zypper's log when running
with ZYPP_FULLLOG=1.

Also remove the superfluous CPP warning. ifstream.getline() throws an
exception in case something goes wrong while reading the file.

Signed-off-by: Valentin Rothberg <vrothberg@suse.com>