rename: handle EXDEV error on OverlayFS
The rename(2) systemcall doesn't succeed across devices, which is likely
to happen when using the overlay storage driver in Docker when trying to
rename between the upper and the lower layer. In such case, errno is
set to EXDEV.

Extend PathInfo::rename to handle the EXDEV errno case and fallback to
using mv(1), which is explicitly mentioned in the kernel docs to deal
correctly with OverlayFS issues.

Fixes: https://github.com/openSUSE/docker-containers/issues/57
Signed-off-by: Valentin Rothberg <vrothberg@suse.com>