Improve Italian traslation of the "breaking dependencies" message.
In Italian, "to break" can indeed be translated with "interrompere". But
that depends on the context, and this one here is really not the case.

Adopt a sentence which, albeit not a literal translation, sounds a lot
more natural and easier to understand.

FTR, this was reported (by me) as a bug here:
 https://bugzilla.opensuse.org/show_bug.cgi?id=1173529