Add tools/zypper-donload-stats script
Grep downloads from a zypper.log file. Show the request URLs and embed
the mirror stats in case of metalink downloads. Lines starting with #0..#N
show the number of parallel connections and the number of byte received.

Ment to be an aid until zypper is able to get those stats from libzypp.