Fix purge-kernels is broken in Leap 15.3 (bsc#1185325)
Leap 15.3 introduces a new kernel package called kernel-flavour-extra,
which contain kmp's. Currently kmp's are detected by name ".*-kmp(-.*)?"
but this does not work which those new packages. This patch fixes the
problem by checking packages for kmod(*) and ksym(*) provides and only
falls back to name checking if the package in question does not provide
one of those.