Fix purge-kernels fails with kernels from Kernel:HEAD (bsc#1187738)
There recently was a change in the kernel package naming scheme in regards
to rc kernels. Since kernel upstream uses characters in the version
that are not allowed in rpm versions a "-rc" was previously replaced with
".rc" which broke sorting by version, to fix this issue it was replaced with "~rc",
which unfortunately broke the purge-kernels logic.
This patch makes sure purge-kernel does apply the same conversion.