Show signature details in a separate table
There are at least two pieces of information that we need
to display for each signature: the "familiarity" of the
signature (unknown, known, or trusted) and the key ID of
the signature.

This data might be easier to display to the user in a
tabular format, as in this commit.