Remove Medianetwork and dependend code
This commit removes the MediaNetwork tech preview and all related code.
First reason for this is that MediaNetwork was just meant as
a way to test the new CURL based downloader
and second: since the Provide API is going to completely replace the
current media backend it would be extra work to ensure that changes
on the Downloader do not break MediaNetwork.