Provide "removeptf" command (bsc#1203249)
A remove command which prefers replacing dependant packages to removing
them as well.

A PTF is typically removed as soon as the fix it provides is applied to
the latest official update of the dependant packages. But you don't want
the dependant packages to be removed together with the PTF, which is what
the remove command would do. The removeptf command however will aim to
replace the dependant packages by their official update versions.