Do not unconditionally release a medium if provideFile failed (bsc#1211661)
MediaSetAccess continues and re-uses an attached medium as long as files
are provided successfully. Unconditionally releasing the medium after
a file could not be provided is not needed. We can continue with the
next file, if there is one.

We can leave it to the exception handling. If the MediaSetAccess goes out
of scope, it's released anyway. If the exception is handled earlier, we
can continue to use it.

Only in case we interactively request a new medium from the user, the
medium needs to be discarded. But this case calls MediaManager::releaseAll
anyway.