More consistency as to when blank lines are printed.
This just makes a some blank lines that were printed to seperate info
output not print when the associated info isn't printed.

It also makes some blank lines print (when not using --quiet) to
seperate output from the initial messages (e.g. "Loading repository
data...").

Some commands where already doing the above two, whilst some weren't.

The repository priority summary is also no longer printed in --quiet
mode. (before the summary was, but not the "Repository priorities in
effect" heading, which made little sense).

Similarly, the hint in list-patches to use --with-optional is now hidden
even in quiet mode.

Note: it would probably be neater to modify zypper.out().gap() to take a
verbosity paramater, but that's defined in another repo, so I couldn't
be bothered.