Add an --ids-only flag to various subcommands.
Most of the subcommands that print a table of information now have an
--ids-only flag, that will print only the ids/names of the table
(this implies --quiet). The following subcommands are affected:
	list-updates
	list-patches
	patches
	patterns
	packages
	repos
	locales
	search
	list-patches
	products
	services
Note that the 'id' is usually the name column, except:
	for repos and services it is the 'alias'
	for products it is the "internal name"
	for locales it is the "code"
	for patches it is the "patch"
The --ids-only flag has no effect when:
	outputing xml
	using repos --export
	using repos with a list of arguments  (as you already know the ids)
The --ids-only flag is not added to:
	patch-check (as that doesn't actually print patch names, just counts)
	white-provides (as that's obsolete)
	info (as the point of that is to get more information)
No documentation or tests cases have been added yet, but I have tested
that each case in the code that prints one of these tables, now prints
only a list of ids (with no blank lines).