info: Enhance pattern content table (bsc#1158038)
Alternatives (multiple packages providing the same requirement)
are now listed as a single entry in the content table. The entry
shows either the installed package which satisfies the requirement
or the requirement itself as type 'Provides'.

Listing all potential alternatives was miss leading, especially if
the alternatives were mutual exclusive. It looked like an installed
pattern had not-installed requirements and it was not possible to install
all requirements at the same time.